Moved a lot of files to try to better organize the project's folder.
Added some properties in the airplane to try to have the environment
more omniscient.

Continued to work on the interface of the simulation.
Added a delegate in order for the environment to ask for graphical
changes.

Renamed the type of message type, to reflect that they are in fact
codes.

Changed the type of application in order to stop it when not opened.