Prepared the creation of the simulation, with the necessary methods to create the different agents inside it. Prepared two methods to start and stop the simulation.
Modified the zones that are now aware of their controller's name, and
if it is a normal zone or one with an airport.

In the display delegate, added some methods to add bulk airplanes when
starting a simulation for example.

The main view controller now has a specific view to display all the
airplanes, so that it can be flushed. The main view controllers also
implements the different methods from the display delegate.