Prevented the application from being displayed in portrait mode, so that it is easier to implement the different views.
Changed the delegate used by the agents to specify methods to implement,
and added a method invoked when the simulation is starting.

Implemented the latter method in the airplane agent, and added some
content to the message analyzer.

In the environment added some methods to request the display to be
updated with new content (mostly the positions of the airplanes).

Changed the main view controller relationships with the nib file, so
that the different layers composing the interface can be reset
separately when restarting a game.