In the environment, added some methods to perform the interface changes through the delegate on the main thread.
Changed the hierarchy between the app delegate and the view controller:
the view controller is created without a nib file.

Changed the type of the keys in the dictionary used to store the views
associated to each airplane (issue with NSCopying, and copyWithZone:
method). Needed probably is a method to ensure the agent names are
uniques.

Tmp printing of a black square instead of the picture.
Needed is also a way to trigger the update of the positions of the
airplane (maybe through messages).