First working representation of all the elements of the simulation.
Added a custom view to draw the borders of the zones.
Added some initial elements.
Modified the airports so that they hold their name and location.
Changed the way the interface painting is called from the environment,
to combine all the display calls in one method.

Added some messages codes.

The controllers holds different layers containing different aspects of
the view.